requirements.txt
numpy
opencv-python
dlib
Pillow
firebase-admin
face-recognition-models